backend/requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pymupdf>=1.24.0
chromadb>=0.5.0
python-dotenv>=1.0.0
